Make all changes after reviewer comments
